More minor fixes
